Refactor: Extract OrchestratorShowcase SVG and detail panel into subcomponents (Fixes #1421)
- Extracted SVG diagram to OrchestratorDiagram
- Extracted detail panel cards to OrchestratorDetailPanel
- Added descriptive aria-label to SVG for accessibility
- Extracted types and data to separate files
- Reduced OrchestratorShowcase length to under 100 lines
